requirements.txt
# Use this file to create an environment using:
# $ conda create --name env_name -c bioconda  --file <this file>

# REQUIRED: DO NOT CHANGE
python>=3.5
graphviz>=2.38.0
snakemake>=3.4.2
click
Sphinx
coverage
awscli
flake8
# python-dotenv>=0.5.1
munch
pip
ruamel.yaml

# Packages specific to your pipeline
synapseclient>=1.5.1
jupyter
jupyter_nbextensions_configurator
widgetsnbextension
jupyter_contrib_nbextensions
pandoc

ipython
seaborn
pandas
numexpr
numpy
xlrd
xlwt
